% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rbstwwk-package.R
\docType{package}
\name{rbstwwk-package}
\alias{rbstwwk}
\alias{rbstwwk-package}
\title{rbstwwk: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: Stefan Moog \email{stefan.moog@prognos.com}

}
\keyword{internal}
