% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rwwk-package.R
\docType{package}
\name{rwwk-package}
\alias{rwwk}
\alias{rwwk-package}
\title{rwwk: Access the Wegweiser Kommune (WWK) Open Data API}
\description{
Provides convenient access to the open municipal datasets from Wegweiser Kommune.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/trekonom/rwwk}
  \item Report bugs at \url{https://github.com/trekonom/rwwk/issues}
}

}
\author{
\strong{Maintainer}: Stefan Moog \email{moogs@gmx.de}

}
\keyword{internal}
