% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wwk-export-indicator.R
\name{wwk-export-indicator}
\alias{wwk-export-indicator}
\alias{wwk_export_indicator}
\title{Export Indicator Meta Data}
\usage{
wwk_export_indicator(
  indicator = NULL,
  format = "csv",
  charset = "UTF-8",
  verbose = FALSE
)
}
\arguments{
\item{indicator}{a character vector of friendly indicator URLs.}

\item{format}{format in which data shgould be returned. One of \code{csv} or \code{json}.}

\item{charset}{encoding to be used. Defaults to \code{UTF-8}.}

\item{verbose}{be verbose}
}
\value{
a character (csv) or a list (json)
}
\description{
Exports the meta data for the specified indicators in the desired format.
}
\examples{
\dontrun{
# Search for one indicator
wwk_export_indicator(
  indicator = "geburten"
)
# Multiple indiciators
wwk_export_indicator(
  indicator = c("geburten", "sterbefaelle")
)
}
}
