% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wwk-export.R
\name{wwk-export}
\alias{wwk-export}
\alias{wwk_export}
\title{Export data}
\usage{
wwk_export(
  indicator = NULL,
  region = NULL,
  year = NULL,
  format = "csv",
  charset = "UTF-8",
  verbose = FALSE
)
}
\arguments{
\item{indicator}{a character vector fiendly indicator URLs}

\item{region}{a character vector of fiendly region URLs}

\item{year}{a character. Years for which data should be retrieved.}

\item{format}{format in which data shgould be returned. One of \code{csv} or \code{json}.}

\item{charset}{encoding to be used. Defaults to \code{UTF-8}.}

\item{verbose}{be verbose}
}
\value{
a character (csv) or a list (json)
}
\description{
Exports the statistical data for the specified 'Friendly URL' in the desired
format.
}
\examples{
\dontrun{
# Search for indicators
wwk_export(
  indicator = "geburten",
  region = "koeln",
  year = "2024"
)
# Multiple indiciators, regions and years
wwk_export(
  indicator = c("geburten", "sterbefaelle"),
  region = c("koeln", "muenchen"),
  year = "2000-2024"
)
}
}
