% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wwk-get.R
\name{wwk-get}
\alias{wwk-get}
\alias{wwk_get}
\alias{wwk_get_indicator}
\alias{wwk_get_topic}
\alias{wwk_get_region}
\title{List available data topics, indicators and regions}
\usage{
wwk_get(what, friendly_url, verbose = FALSE)

wwk_get_indicator(indicator, verbose = FALSE)

wwk_get_topic(topic, verbose = FALSE)

wwk_get_region(region, verbose = FALSE)
}
\arguments{
\item{what}{a character. One of \code{"indicator"}, \code{"topic"} or \code{"region"}}

\item{friendly_url, indicator, topic, region}{friendly url of indicator, topic or region}

\item{verbose}{be verbose}
}
\value{
a tibble
}
\description{
Retrieves a list of topics, indicators and regions from the
Wegweiser Kommune Open Data API.
}
\examples{
\dontrun{
# Get meta data on an indicator
wwk_list_indicator("geburten")
# Get meta data on a region
wwk_get_region("koeln")
# Get meta data on a topic
wwk_get_topic("anteil-der-altersgruppen")
}
}
