% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wwk-list.R
\name{wwk-list}
\alias{wwk-list}
\alias{wwk_list}
\alias{wwk_list_indicator}
\alias{wwk_list_topic}
\alias{wwk_list_region}
\title{List available data topics, indicators and regions}
\usage{
wwk_list(what, max = 10, ..., verbose = FALSE)

wwk_list_indicator(max = 10, ..., verbose = FALSE)

wwk_list_topic(max = 10, ..., verbose = FALSE)

wwk_list_region(max = 10, ..., verbose = FALSE)
}
\arguments{
\item{what}{a character. One of \code{"indicator"}, \code{"topic"} or \code{"region"}}

\item{max}{maximum number of results to return}

\item{...}{additional parameters}

\item{verbose}{be verbose}
}
\value{
A data.frame.
}
\description{
Retrieves a list of topics, indicators and regions from the
Wegweiser Kommune Open Data API.
}
\examples{
\dontrun{
wwk_list_indicator()
# Set the maximum number of results
wwk_list_indicator(max = 5)
# Search for indicators
wwk_list_indicator(search = "Frauen")
# Exclude indicators by ID
wwk_list_indicator(exclude = c(255, 273))
}
}
