% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wwk-suggest.R
\name{wwk-suggest}
\alias{wwk-suggest}
\alias{wwk_suggest}
\alias{wwk_suggest_demographic_type}
\alias{wwk_suggest_gkz}
\alias{wwk_suggest_population_range}
\alias{wwk_suggest_region_type}
\title{Suggest}
\usage{
wwk_suggest(
  what,
  max = 10,
  exclude = NULL,
  search = NULL,
  ...,
  verbose = FALSE
)

wwk_suggest_demographic_type(
  max = 10,
  exclude = NULL,
  search = NULL,
  verbose = FALSE
)

wwk_suggest_gkz(
  max = 10,
  exclude = NULL,
  search = NULL,
  include_ars = TRUE,
  verbose = FALSE
)

wwk_suggest_population_range(
  max = 10,
  exclude = NULL,
  search = NULL,
  verbose = FALSE
)

wwk_suggest_region_type(
  max = 10,
  exclude = NULL,
  search = NULL,
  verbose = FALSE
)
}
\arguments{
\item{what}{a character. One of \code{"indicator"}, \code{"topic"} or \code{"region"}}

\item{max}{maximum number of results to return}

\item{exclude}{a character. Optional list of IDs of indicators, topics or
regions to exclude.}

\item{search}{a character. Optional search text to look for indicators,
topics or regions.}

\item{...}{additional parameters}

\item{verbose}{be verbose}

\item{include_ars}{a boolean. Include 12-digit ARS in the search.}
}
\value{
A data.frame.
}
\description{
Get a list of suggested values to filter by demographic type,
regional key, population range or region type.
}
\examples{
\dontrun{
wwk_list_indicator()
# Set the maximum number of results
wwk_list_indicator(max = 5)
# Search for indicators
wwk_list_indicator(search = "Frauen")
# Exclude indicators by ID
wwk_list_indicator(exclude = c(255, 273))
}
}
