% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wwk-api.R
\name{wwk_api}
\alias{wwk_api}
\title{WWK API}
\usage{
wwk_api()
}
\value{
invisible returns API specs as a list
}
\description{
A convenience function to browse the API specs using \code{listviewer} (if installed).
}
\examples{
wwk_api()
}
