% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_metadata.R
\name{wwk_get_metadata}
\alias{wwk_get_metadata}
\title{Get metadata for a specific indicator}
\usage{
wwk_get_metadata(indicator_id)
}
\arguments{
\item{indicator_id}{The ID of the indicator.}
}
\value{
A list of metadata.
}
\description{
Get metadata for a specific indicator
}
