% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_regions.R
\name{wwk_get_regions}
\alias{wwk_get_regions}
\title{Get list of regions}
\usage{
wwk_get_regions()
}
\value{
A data.frame of region IDs and names.
}
\description{
Returns a data.frame of available regions (municipalities) with their IDs.
}
