% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wwk-read-csv.R
\name{wwk_read_csv}
\alias{wwk_read_csv}
\title{Read WWK CSV files}
\usage{
wwk_read_csv(file, skip = 2, n_max = Inf, show_col_types = FALSE)
}
\arguments{
\item{file}{Either a path to a file, a connection, or literal data
(either a single string or a raw vector).

Files ending in \code{.gz}, \code{.bz2}, \code{.xz}, or \code{.zip} will
be automatically uncompressed. Files starting with \verb{http://},
\verb{https://}, \verb{ftp://}, or \verb{ftps://} will be automatically
downloaded. Remote gz files can also be automatically downloaded and
decompressed.

Literal data is most useful for examples and tests. To be recognised as
literal data, the input must be either wrapped with \code{I()}, be a string
containing at least one new line, or be a vector containing at least one
string with a new line.

Using a value of \code{\link[readr:clipboard]{clipboard()}} will read from the system clipboard.}

\item{skip}{Number of lines to skip before reading data. If \code{comment} is
supplied any commented lines are ignored \emph{after} skipping.}

\item{n_max}{Maximum number of lines to read.}

\item{show_col_types}{If \code{FALSE}, do not show the guessed column types. If
\code{TRUE} always show the column types, even if they are supplied. If \code{NULL}
(the default) only show the column types if they are not explicitly supplied
by the \code{col_types} argument.}
}
\value{
a tibble
}
\description{
Function to return WWK data retrieved in csv format as a tidy tibble.
}
\details{
The CSV files returned by the WWK API cannot be used directly for further analyses, as they contain:
\enumerate{
\item Lines with metadata at the beginning and end of the data set.
\itemize{
\item The first line contains the names of the indicators.
\item The second line contains the names of the regions.
\item Additional meta data is provided at the end.
}
\item The data is returned in wide format:
\itemize{
\item Each row corresponds to an indicator.
\item The first column contains the name of the indicator.
\item All further columns contain the data for a region-year pair.
}
}
}
\examples{
\dontrun{
csv <- wwk_export(
  indikator = c("geburten", "sterbefaelle"),
  region = c("koeln", "muenchen")
)
wwk_read_csv(csv)
}
}
