% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rwwk-package.R
\docType{package}
\name{wwkapir-package}
\alias{wwkapir}
\alias{wwkapir-package}
\title{wwkapir: Access the Wegweiser Kommune (WWK) Open Data API}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Provides convenient access to the open municipal datasets from Wegweiser Kommune.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/trekonom/wwkapir}
  \item \url{https://trekonom.github.io/wwkapir/}
  \item Report bugs at \url{https://github.com/trekonom/wwkapir/issues}
}

}
\author{
\strong{Maintainer}: Stefan Moog \email{moogs@gmx.de}

}
\keyword{internal}
