% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rbstwwk-package.R
\docType{package}
\name{rbstwwk-package}
\alias{rbstwwk}
\alias{rbstwwk-package}
\title{rbstwwk: Access Wegweiser Kommune Open Data API}
\description{
Provides convenient access to the open municipal datasets from Wegweiser Kommune.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/trekonom/rbstwwk}
  \item Report bugs at \url{https://github.com/trekonom/rbstwwk/issues}
}

}
\author{
\strong{Maintainer}: Stefan Moog \email{moogs@gmx.de}

}
\keyword{internal}
